call api product

+++ src/main/java/vn/com/shop/dto/product/ProductVariantResponseDTO.java
@@ -4,14 +4,13 @@ import lombok.Getter;
 import lombok.Setter;
 import vn.com.shop.dto.BaseResponseDTO;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
 public class ProductVariantResponseDTO extends BaseResponseDTO {
     private String sku;
-    private BigDecimal price;
-    private BigDecimal discountPrice;
+    private Double price;
+    private Double discountPrice;
     private Integer stock;
     private Integer reservedStock;
     private String size;
